third_party/WebKit/LayoutTests/fast/viewport/viewport-77-expected.txt

CONSOLE MESSAGE: line 3: The key ";" is not recognized and ignored.
CONSOLE MESSAGE: line 3: The value "yes;" for key "initial-scale" is invalid, and has been ignored.
CONSOLE MESSAGE: line 3: Note that ';' is not a key-value pair separator. The list should be comma-separated.
ALERT: viewport size 1280x1408 scale 0.25 with limits [0.25, 5] and userScalable true
